+++ TracageAlmentaireWeb/Controllers/ApiControllers/OrganizationsController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Web.Http;
 using Tracage.Models;
+using TracageAlmentaireWeb.BL.RESTSecurityLayer;
 using TracageAlmentaireWeb.DAL;
 
 namespace TracageAlmentaireWeb.Controllers.ApiControllers
@@ -9,38 +10,75 @@ namespace TracageAlmentaireWeb.Controllers.ApiControllers
     {
 
         Mapper mapper = new Mapper("FTDb");
+        private APIKeyMessageHandler keyHandler = new APIKeyMessageHandler();
 
         [Route("api/Organizations")]
-        public IEnumerable<Organization> Get()
+        public IHttpActionResult Get()
         {
-            return mapper.GetOrganizations();
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                IEnumerable<Address> ad = mapper.GetAddresses();
+
+                if (ad != null)
+                {
+                    return Ok(mapper.GetOrganizations());
+                }
+                else
+                {
+                    return NotFound();
+                }
+            }
+            return new ForbiddenActionResult(Request, "access denied");
+
         }
 
         [Route("api/Organizations/{id}")]
         public IHttpActionResult Get(long id)
         {
-            var result = mapper.GetOrganization(id);
-            if (result != null)
-            {
-                return Ok(result);
-            }
 
-            return NotFound();
+            IHttpActionResult reponse;
+
+            if (!keyHandler.CheckApiKey(this.Request))
+            {
+                return new ForbiddenActionResult(Request, "access denied");
+            }
+            else
+            {
+                var result = mapper.GetOrganization(id);
+                if (result != null)
+                {
+                    reponse = Ok(result);
+                }
+                else
+                {
+                    reponse = NotFound();
+                }
+            }
+            return reponse;
         }
 
         public void Post(Organization data)
         {
-            mapper.CreateOrganization(data);
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                mapper.CreateOrganization(data);
+            }
         }
 
         public void Put(long id, Organization data)
         {
-            mapper.UpdateOrganization(id,data);
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                mapper.UpdateOrganization(id, data);
+            }
         }
 
         public void Delete(long id)
         {
-            mapper.DeleteOrganization(id);
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                mapper.DeleteOrganization(id);
+            }
         }
 
     }
