+++ TracageAlmentaireWeb/Controllers/ApiControllers/RolesController.cs
@@ -3,6 +3,7 @@ using System.Security.Authentication.ExtendedProtection;
 using System.Web.Http;
 using Microsoft.Ajax.Utilities;
 using Tracage.Models;
+using TracageAlmentaireWeb.BL.RESTSecurityLayer;
 using TracageAlmentaireWeb.DAL;
 
 namespace TracageAlmentaireWeb.Controllers.ApiControllers
@@ -10,40 +11,73 @@ namespace TracageAlmentaireWeb.Controllers.ApiControllers
     public class RolesController : ApiController
     {
 
-       Mapper mapper =new Mapper("FTDb");
-        
+        Mapper mapper = new Mapper("FTDb");
+        private APIKeyMessageHandler keyHandler = new APIKeyMessageHandler();
+
+
         [Route("api/Roles")]
-        public IEnumerable<Role> Get()
+        public IHttpActionResult Get()
         {
-            return mapper.GetRoles();
+           
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                return Ok(mapper.GetRoles());
+            }
+            else
+            {
+                return new ForbiddenActionResult(Request, "access denied");
+            }
+
         }
 
         [Route("api/Roles/{id}")]
         public IHttpActionResult Get(long id)
         {
-            var result = mapper.GetRole(id);
-            if (result != null)
-            {
-                return Ok(result);
-            }
 
-            return NotFound();
+            if (!keyHandler.CheckApiKey(this.Request))
+            {
+                return new ForbiddenActionResult(Request, "access denied");
+            }
+            else
+            {
+                var result = mapper.GetRole(id);
+                if (result != null)
+                {
+                    return Ok(result);
+                }
+                else
+                {
+                    return NotFound();
+                }
+            }
+            
         }
 
-       
+
         public void Post(Role data)
         {
-            mapper.CreateRole(data);
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                mapper.CreateRole(data);
+            }
         }
 
         public void Put(long id, Role data)
         {
-            mapper.UpdateRole(id,data);
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                mapper.UpdateRole(id, data);
+            }
+                
         }
 
         public void Delete(long id)
         {
-            mapper.DeleteRole(id);
+            if (keyHandler.CheckApiKey(this.Request))
+            {
+                mapper.DeleteRole(id);
+            }
+                
         }
 
     }
